extended user password helper utility class to include password strength validation.
